+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * added support for passing weighted igraph objects
 
+* improved handling of sparse matrices
+
 * bug fixes to ensure same results from matrix and igraph methods
 
 # leiden 0.3.1